Reviewed resources.
Changed travel budget from Silesia, which was based on 0.65 FTE
at some point, and is now back at 0.5 FTE (24 months out of 48
runtime). This also changes the total (a bit more in the change due to
the 25% overheads).

Allowed to use the 180k for services other than Binder.

Mentioned that we will use meeting rooms of our organisations where possible.

Otherwise just stylistic changes and presentation.